{{ cookiecutter.repo_name }}/requirements.txt
# local package
-e .

# external requirements
pytorch=1.9.0
torchvision
torchaudio
cudatoolkit=11.1
tensorboard
numpy
scipy
scikit-learn
matplotlib
seaborn
pandas
click
absl-py
pyyaml
jupyterlab
tqdm
sphinx
awscli
python-dotenv
duecredit
{% if cookiecutter.python_interpreter != 'python3' %}

# backwards compatibility
future
pathlib2
{% endif %}
